common filters: pt1FilterGain must accept uint16_t (#5513)
* all places that call pt1FilterGain use lpfHz which is of type uint16_t. I have
  not confirmed yet but believe that when lpfHz were greater than 255 we would
  have seen data loss / overflow / wrap around. TBC

+++ src/main/common/filter.c
@@ -42,7 +42,7 @@ FAST_CODE float nullFilterApply(filter_t *filter, float input)
 
 // PT1 Low Pass filter
 
-float pt1FilterGain(uint8_t f_cut, float dT)
+float pt1FilterGain(uint16_t f_cut, float dT)
 {
     float RC = 1 / ( 2 * M_PI_FLOAT * f_cut);
     return dT / (RC + dT);
